ms4/requirements.txt
fastapi
uvicorn
kafka-python
python-dotenv
pandas
sqlalchemy

psycopg2-binary